python/oolama/requirements.txt
langchain-openai 
langchain_core
python_dotenv
streamlit
langchain_community
